Kattis - everywhere

+++ everywhere/src/main.rs
@@ -4,12 +4,13 @@ use std::collections::HashSet;
 fn main() {
     let mut input = String::new();
     io::stdin().read_line(&mut input).unwrap();
-    let t: u32 = input.trim().parse().unwrap();
+
+    let t: usize = input.trim().parse().unwrap();
 
     for _ in 0..t {
         input.clear();
         io::stdin().read_line(&mut input).unwrap();
-        let n: u32 = input.trim().parse().unwrap();
+        let n: usize = input.trim().parse().unwrap();
 
         let mut cities = HashSet::new();
